+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - Allow Multiple Notifications
+
+* New: Add ability to trigger multiple notifications on single device
+
+
 ## [0.1.0] - Initial Commit and Public Release
 
 First public release